tidy up names and tests
