+++ api/tests/utils.py
@@ -1,4 +1,5 @@
-from app.main import app
 from fastapi.testclient import TestClient
+
+from app.main import app
 
 test_client = TestClient(app)
